requirements_no_version.txt
asgiref
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
ffmpeg-python
matplotlib
numpy
opencv-python
Pillow
scipy
sqlparse
tqdm
tzdata
decord
paddlex[ocr]
paddlenlp
# git+https://github.com/facebookresearch/sam3.git
